requirements.txt
wheel
tqdm
pytorch_lightning
black
ruff
hydra-core
jaxtyping
beartype
wandb
einops
colorama
scikit-image
colorspacious
matplotlib
moviepy==1.0.3
imageio
git+https://github.com/dcharatan/diff-gaussian-rasterization-modified
timm
dacite
lpips
e3nn
plyfile
tabulate
svg.py
opencv-python==4.6.0.66
sk-video
torchmetrics
